add rent skill of clan

+++ Data/Clan/ClanSkillTemplate.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Gopet.Data.Clan
 {
-    public record ClanSkillTemplate(int id, string name, string description, long expire, int[] moneyType, int[] price)
+    public record ClanSkillTemplate(int id, string name, string description, long expire, sbyte[] moneyType, int[] price)
     {
         public ClanSkillLvlTemplate[] clanSkillLvlTemplates { get; set; }
     }
